+++ IsaacSocket-DLL/CMakeLists.txt
@@ -6,9 +6,13 @@ endif()
 set(CMAKE_CXX_STANDARD 20)
 
 set(CMAKE_TOOLCHAIN_FILE ${CMAKE_CURRENT_LIST_DIR}/mingw-toolchain.cmake)
+if (CMAKE_CURRENT_SOURCE_DIR EQUAL CMAKE_CURRENT_BINARY_DIR)
+    message(FATAL_ERROR "cannot use same source and binary dir!")
+endif()
 
-project(IsaacSocket LANGUAGES CXX)
+project(IsaacSocket LANGUAGES CXX ASM)
 
-file(GLOB files *.cpp *.hpp *.h)
-add_library(IsaacSocket SHARED ${files})
+file(GLOB files *.cpp *.hpp *.h *.S)
+add_library(IsaacSocket SHARED)
+target_sources(IsaacSocket PRIVATE ${files})
 target_include_directories(IsaacSocket PRIVATE include)
