+++ IsaacSocket-DLL/isaac.hpp
@@ -1,12 +1,9 @@
 ﻿#pragma once
 #include "pch.h"
-
-using std::string;
-using std::vector;
+#include "utils.hpp"
+#include "gnucompat.hpp"
 
 namespace isaac {
-
-#define FIELD(address,type,name) struct{char pad_##address[address];type name;}
 
 	typedef struct lua_State lua_State;
 
@@ -61,9 +58,8 @@ namespace isaac {
 			//0:关闭 2:开启 4:正在关闭 小于0:强制暂停
 			FIELD(0x20, int32_t, state);
 			FIELD(0x30, float, offsetY);
-			FIELD(0x3C, string, consoleText);
+			FIELD(0x3C, gnucompat::string, consoleText);
 		};
-		~Console() {}
 	};
 
 	// 主动道具类
@@ -116,9 +112,8 @@ namespace isaac {
 			FIELD(0x132C, Weapon*, weapon);
 			FIELD(0x14C4, Active, actives[4]);
 			FIELD(0x1745, uint8_t, canShoot);
-			FIELD(0x1D0C, vector<Passive>, passives);
+			FIELD(0x1D0C, gnucompat::vector<Passive>, passives);
 		};
-		~Player() {}
 	};
 
 	//种子
@@ -139,7 +134,7 @@ namespace isaac {
 	struct GlitchedItem
 	{
 		char pad_0x2C[0x2C];
-		vector<GlitchedEvent*> glitchedEvents;
+		gnucompat::vector<GlitchedEvent*> glitchedEvents;
 	};
 
 	//GLFW窗口类
@@ -159,14 +154,13 @@ namespace isaac {
 	{
 		union
 		{
-			FIELD(0x1BA50, vector<Player*>, players);
+			FIELD(0x1BA50, gnucompat::vector<Player*>, players);
 			FIELD(0x1BB04, Seeds, seeds);
 			FIELD(0x1BB60, Console, console);
 			FIELD(0x10203C, PauseMenu, pauseMenu);
 			FIELD(0x1C3164, uint32_t, debugFlag);
-			FIELD(0x283C20, vector<GlitchedItem*>, glitchedItems);
+			FIELD(0x283C20, gnucompat::vector<GlitchedItem*>, glitchedItems);
 		};
-		~Game() {}
 	};
 
 	// 以撒主模块镜像类
@@ -181,5 +175,4 @@ namespace isaac {
 			FIELD(0x7FD688, MainMenu*, mainMenu);
 		};
 	};
-#undef FIELD
 }
