+++ IsaacSocket-DLL/mingw-toolchain.cmake
@@ -11,13 +11,11 @@ set(CMAKE_SYSTEM_NAME Windows)
 set(TOOLCHAIN_PREFIX i686-w64-mingw32)
 
 # cross compilers to use for C and C++
-#set(CMAKE_C_COMPILER ${TOOLCHAIN_PREFIX}-gcc)
-#set(CMAKE_CXX_COMPILER ${TOOLCHAIN_PREFIX}-g++)
 set(CMAKE_C_COMPILER ${TOOLCHAIN_PREFIX}-gcc)
 set(CMAKE_CXX_COMPILER ${TOOLCHAIN_PREFIX}-g++)
 set(CMAKE_RC_COMPILER ${TOOLCHAIN_PREFIX}-windres)
 
-# target environment on the build host system
+# target environment on the build host system, the 13.1.0 may vary from version to version
 set(CMAKE_FIND_ROOT_PATH /usr/${TOOLCHAIN_PREFIX} /usr/lib/gcc/${TOOLCHAIN_PREFIX}/13.1.0)
